requirements.txt
python-telegram-bot[job-queue]>=20.5
python-dotenv
aiosqlite
httpx
pytz
